requirements.txt
google-generativeai
feedparser
requests
python-dotenv
schedule
markdown
beautifulsoup4 # For potential future content extraction, good to have
# Add sendgrid if preferred over smtplib
# sendgrid 